requirements-nim-gateway.txt
fastapi
uvicorn
litellm
pyyaml
